Move HTTP code into own subdir
